Prevent TF test to execute in parallel.
